+ Fixed a compilation warning
git-svn-id: https://scm.blissgarden.org/svn/lima/trunk@91 45c4679d-1946-429d-baad-37912b19538b

+++ camera/common/espia/src/EspiaAcq.cpp
@@ -113,7 +113,8 @@ void Acq::lastFrameCallback(struct espia_cb_data *cb_data)
 	} else {
 		m_last_frame_info = cb_finfo;
 		bool endless = (m_nb_frames == 0);
-		if (!endless && (cb_finfo.acq_frame_nr == m_nb_frames - 1))
+		if (!endless && 
+		    (cb_finfo.acq_frame_nr == (unsigned long)m_nb_frames - 1))
 			m_started = false;
 	}
 }
